Create Lesson 20 Main Class.java
The main class for Project… How Far To The Line?
Here is how things will be organized in the class:
x Call the class DistToLine. 
x Create static double state variables A, B, and C. 
x Create the signature of the static method getDist. It will receive doubles a and b. It will 
return a double representing the calculated distance.
x In the body of this method, implement the distance formula above and return that value.