Create Lesson 20 Tester.java
Create a Tester class as part of this project that will:
x Set the static state variables A, B, and C with the corresponding values of the desired line 
entered from the keyboard. 
x Call the method getDist and pass as arguments the coordinates of the desired point 
entered from the keyboard. 
x Print the returned double as the distance from the point to the line.