Lesson 15 Output
Create a Tester class that has a main( ) method. In that method you should input from the 
keyboard the amount (1000) of money initially to be put into the account (via the constructor) 
along with the name of the person to whom the account belongs. 
1. Use these two pieces of data to create a new BankAccount object called myAccount. 
2. Call the deposit method to deposit 505.22. 
3. Print the balance state variable.
4. Call the withdraw method to withdraw 100. 
5. Print the remaining balance in this form:
The Sally Jones account balance is, $1405.22