Lesson 15 Info
Project … Overdrawn at the Bank
Create a class called BankAccount. It should have the following properties: 
1. Two state variables:
double balance… This is how much money is currently in the account. 
String name…The name of the person owning the account. 
2. Constructor should accept two parameters. 
a. One should be a double variable that is used to initialize the state variable, 
balance. 
b. The other should be a String that is used to initialize the state variable, name. 
3. Two methods: 
a. deposit…returns nothing…accepts a double that is the amount of money being 
deposited. It is added to the balance to produce a new balance.
b. withdraw…returns nothing…accepts a double that is the amount of money 
being taken out of the account. It is subtracted from the balance to produce a 
new balance. 
Create a Tester class that has a main( ) method. In that method you should input from the 
keyboard the amount (1000) of money initially to be put into the account (via the constructor) 
along with the name of the person to whom the account belongs. 
1. Use these two pieces of data to create a new BankAccount object called myAccount. 
2. Call the deposit method to deposit 505.22. 
3. Print the balance state variable.
4. Call the withdraw method to withdraw 100. 
5. Print the remaining balance in this form:
The Sally Jones account balance is, $1405.22